Further code refactoring / Fix export types
